+++ modules/post/components/creation/meetingNameInput/MeetingNameInput.tsx
@@ -6,7 +6,7 @@ import TextLengthCounter from '@shared/components/textLengthCounter/TextLengthCo
 import Input from '@shared/components/input/Input';
 import { truncateText } from '@shared/utils/utils';
 import ScreenLayout from '@shared/components/layout/ScreenLayout';
-import DividerWrapper from '@post/components/detail/DividerWrapper';
+import DividerWrapper from '@shared/components/layout/DividerWrapper';
 import useStyles from './style';
 
 const TITLE = '모임 이름';
