+++ modules/post/screens/MeetingPostCreator.tsx
@@ -6,7 +6,7 @@ import MeetingNameInput from '@post/components/creation/meetingNameInput/Meeting
 import Uploader from '@post/components/creation/uploader/Uploader';
 import VotingTimeRangePicker from '@post/components/creation/votingTimeRangePicker/VotingTimeRangePicker';
 import ConfirmCancelButton from '@post/components/button/ConfirmCancelButton';
-import DividerWrapper from '@post/components/detail/DividerWrapper';
+import DividerWrapper from '@shared/components/layout/DividerWrapper';
 
 export default function MeetingPostCreator() {
   return (
